Implement the comparison operators >= and <=
This was supposed to be a quick and easy extension. The definitions and
test code are trivial but this turned into a long slog debug fest because
the bug turned out to be quite subtle. On the x86 laptop that I'm using for
this development, an int is 4 bytes long (32-bits) and I'm emulating the
realmode x86 at 16 bits. So for negative numbers the 16-bit sign extended
values don't quite go all the way into 32-bit territory. So my one-liner
solution, for now, is to replace the int used for registers with a short
(which uses 16 bits on the laptop instead of the 32 bits used for an int)

$FORTH >= <=
$BUG emulator needs to use short instead of int for the realmode registers

+++ x86/emulate.c
@@ -49,7 +49,7 @@ typedef struct x86splitreg {
 typedef struct x86register {
 	union {
 		x86splitreg s;
-		int val;
+		short val;
 	};
 } x86register;
